+++ components/layout/Dialog.go
@@ -296,8 +296,25 @@ func (dialog *Dialog) Render() *dom.Element {
 
 }
 
+func (dialog *Dialog) SetContent(component interfaces.Component) {
+
+	dialog.Content = component
+	dialog.Render()
+
+}
+
+func (dialog *Dialog) SetFooter(footer *Footer) {
+
+	dialog.Footer = footer
+	dialog.Render()
+
+}
+
 func (dialog *Dialog) SetTitle(value string) {
+
 	dialog.Title = strings.TrimSpace(value)
+	dialog.Render()
+
 }
 
 func (dialog *Dialog) Show() bool {